Add override to force JSX through comment
**What**
If a file contain the comment `@jsx`, treat it as a JSX file.

**Why**
A mock file for a `.tsx` file must have the `.tsx` file extension for jest to find it.

However, if the mock file does not contain any JSX, the codemod would use the file extension `.ts`.

As a workaround, we introduce the ability to add a `@jsx` comment to the mock file to force the codemod to treat it as a JSX file. This ensures the mock file us
es the `.tsx` extension, even if it doesn't contain any JSX.